requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
pymysql==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Technical Analysis (optional - install TA-Lib separately if needed)
# TA-Lib requires system-level installation
# See: https://ta-lib.github.io/ta-lib-python/
# talib==0.4.28

# Scheduler
APScheduler==3.10.4

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Data Sources
tushare==1.3.7
akshare==1.12.0

# OCR (optional - requires Baidu OCR API)
# aip==4.16.13

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Code Quality
black==23.11.0
flake8==6.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
